+++ src/components/OverflowMenu/OverflowMenu.js
@@ -18,7 +18,7 @@ const Trigger = styled.div`
 const Menu = styled.div`
   display: block;
   margin-top: 4px;
-  background: ${props => props.theme.lightest};
+  background: ${props => props.theme.white};
   min-width: 100px;
   box-shadow: 0 1px 4px rgba(0, 0, 0, .2);
   white-space: nowrap;
@@ -56,11 +56,11 @@ class OverflowMenu extends Component {
     const { appearance, theme } = this.props;
     const colors = {
       primary: theme.primary,
-      secondary: theme.dark,
+      secondary: theme.grey03,
       success: theme.success,
       warning: theme.warning,
-      light: theme.lighter,
-      dark: theme.darkest
+      light: theme.grey06,
+      dark: theme.grey01
     };
     return colors[appearance];
   }