+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### Breaking changes
 - Dropdown and OverflowMenu components now require a `persist` prop to be set to avoid closing dropdown panel after a click inside.
+- New colour names.
 
 ## [2.0.0] - 2018-10-24